Can configure the sanitizer as a standalone object to sanitize arbitrary strings too
